feat: Add carousel of featured movies
This commit adds the a Carousel from a third-party library, we also added a utility function `chunkArray` to the `utils.ts` file. The function takes an array and a size as parameters and returns a new array with the original array chunked into smaller arrays of the specified size. This function can be used to split large arrays into smaller parts to pass a sublist of movies to the carousel since by default it does not split the items per page.

This commit also includes updates to the `types.ts` file, adding a new interface `ApiResponse<T>` to represent the structure of API responses. This interface includes properties for page number, results, total pages, and total results.

Lastly, there is a minor change in the `constants.ts` file, updating the `baseApiUrl` constant to `baseApiUrlMovies` to provide a more descriptive name since we'll add a tv shows api url in the future.