refactor: Add release year to media title in SearchResultGrid
This commit modifies the SearchResultGrid component to display the release year of movies and TV shows in the media title. The release year is extracted from the `release_date` or `first_air_date` property of the media object. This enhancement improves the user experience by providing additional information about the media.